admin page

+++ server/models/models.js
@@ -20,7 +20,7 @@ const Device = sequelize.define("device", {
   id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
   name: { type: DataTypes.STRING, unique: true, allowNull: false },
   price: { type: DataTypes.INTEGER, allowNull: false },
-  rating: { type: DataTypes.INTEGER, defaultValue: 0 },
+  rating: { type: DataTypes.FLOAT, defaultValue: 0 },
   img: { type: DataTypes.STRING, allowNull: false },
 });
 
@@ -36,7 +36,7 @@ const Brand = sequelize.define("brand", {
 
 const Rating = sequelize.define("rating", {
   id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
-  rate: { type: DataTypes.INTEGER, allowNull: false },
+  rate: { type: DataTypes.FLOAT, allowNull: false },
 });
 
 const DeviceInfo = sequelize.define("device_info", {
